requirements.txt
Django~=4.0.2
djangorestframework~=3.13.1
pymongo==3.12.1
djongo

libs~=0.0.10